+++ src/commands/manuals/get.rs
@@ -17,10 +17,10 @@
 
 // This manual page was automatically generated from the mangen.py tool.
 pub const MANUAL: &str = r#"NAME
-    get - display integration code for a core
+    get - display integration code for a unit
 
 SYNOPSIS
-    orbit get [options] <unit>
+    orbit get [options] [<unit>]
 
 DESCRIPTION
     Returns hdl code snippets for the provided design unit to be integrated into
@@ -59,7 +59,8 @@ DESCRIPTION
     
     Exporting the unit's declaration information can be accomplished by using the
     '--json' option. The valid json is unformatted for encouragement to be 
-    processed by other programs.
+    processed by other programs. Retrieve a list of all the unit declarations
+    for the project by omitting '<unit>' while still using the '--json' option.
     
     By default, the code snippets will be displayed in the design unit's native
     hardware description language. To return the code snippets in a particular
@@ -108,4 +109,5 @@ EXAMPLES
     orbit get ram --project mem:2 -csi
     orbit get uart -si --name uart_inst0
     orbit get or_gate -p gates --json
+    orbit get --json
 "#;