+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Add `- unreleased` along the next future version to prevent CI/CD from triggerin
 - Adds support for multiple locations for dependencies using `path` and `version`
 - Allows `version` to be optional when developing locally and specifying `path` for a dependency (`version` is still required for publishing)
 - Allows packages to be set as root for tree when using `orbit tree -e all`
+- Allows all units to be displayed in JSON format using `orbit get --json`
 
 ### Fixes
 - Checks an explicit version (not 0.0.0) is set in the project's manifest when publishing